Added Blockprint Waves
The other change I proposed, didn't work, but I figured out how it works now, and was able to test it on my local machine, and it seemed to work perfectly.
The new block print waves (starting at lines 836 & 1168 ) work the same way as the other ocean textures, and one is simply a inverse of the other.  Just a little eye candy.